requirements.txt
customtkinter>=5.2.0
pyserial>=3.5
opencv-python>=4.8
Pillow>=10.0
pyyaml>=6.0
